backend/requirements.txt
# Dependencias del laboratorio de backtesting Coco Stonks
# Instalar con: pip install -r requirements.txt

requests==2.32.3    # Descarga de ZIPs desde data.binance.vision
pandas==2.2.3       # Manejo de datos y DataFrames
numpy==1.26.4       # Cálculos numéricos (RSI, EMA, Sharpe)
pytz==2024.2        # Conversión de timezone a GMT-3 (Argentina)
PyYAML==6.0.1       # Manejor de archivos de configuración YAML
tqdm==4.66.4        # Barras de progreso para ejecuciones masivas
# AutoLab API — Python dependencies
# Para autolab_api.py (FastAPI) + autolab_brain.py (LLM) + autolab_loop.py

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.8.0

# HTTP client (para llamadas a NVIDIA API)
httpx==0.27.0

# PostgreSQL (Supabase)
psycopg2-binary==2.9.9

# Ya disponibles en el sistema de backtesting (no reinstalar):
# pandas, numpy, sqlite3 (stdlib)
